requirements.txt
tornado==6.4
pydantic==2.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0
python-dotenv==1.0.1